Split up driver kraken_x62 into multiple source files
Move all functionality pertaining to LEDs except the store-functions into its
own object `leds.o`.  Move utility functions into object `util.o`.
