+++ src/util.h
@@ -1,0 +1,8 @@
+#ifndef LEVIATHAN_UTIL_H_INCLUDED
+#define LEVIATHAN_UTIL_H_INCLUDED
+
+#define WORD_LEN_MAX 64
+
+int str_scan_word(const char **buf, char *word);
+
+#endif  /* LEVIATHAN_UTIL_H_INCLUDED */
